refactor: move keybinding parsing from config to keybindings.rs
