+++ crates/omnyssh/src/lib.rs
@@ -6,6 +6,7 @@
 pub mod app;
 pub mod config;
 pub mod event;
+pub mod keybindings;
 pub mod ssh;
 pub mod term_input;
 pub mod ui;
